- Removed local files - Updated titanium.xconfig so the module will build on 7.5.0.GA SDK - Updated the documentation
