--- android/.classpath
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<classpath>
-	<classpathentry kind="src" path="android/src"/>
-	<classpathentry kind="src" path="android/build/.apt_generated"/>
-	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER"/>
-	<classpathentry kind="lib" path="/home/miga/tools/android-sdk-linux/platforms/android-23/android.jar"/>
-	<classpathentry kind="lib" path="/home/miga/tools/android-sdk-linux/add-ons/addon-google_apis-google-23/libs/maps.jar"/>
-	<classpathentry kind="lib" path="/home/miga/.titanium/mobilesdk/linux/6.0.0.GA/android/titanium.jar"/>
-	<classpathentry kind="lib" path="/home/miga/.titanium/mobilesdk/linux/6.0.0.GA/android/kroll-common.jar"/>
-	<classpathentry kind="lib" path="/home/miga/.titanium/mobilesdk/linux/6.0.0.GA/android/kroll-apt.jar"/>
-	<classpathentry kind="output" path="bin"/>
-</classpath>
