--- android/.settings/org.eclipse.jdt.apt.core.prefs
@@ -1,7 +1,0 @@
-#Thu Sep 02 15:18:34 CDT 2010
-eclipse.preferences.version=1
-org.eclipse.jdt.apt.aptEnabled=true
-org.eclipse.jdt.apt.genSrcDir=.apt_generated
-org.eclipse.jdt.apt.reconcileEnabled=true
-
-org.eclipse.jdt.apt.processorOptions/kroll.jsonFile=ti.bluetooth.json
