--- android/.settings/org.eclipse.jdt.core.prefs
@@ -1,3 +1,0 @@
-#Thu Sep 02 15:18:34 CDT 2010
-eclipse.preferences.version=1
-org.eclipse.jdt.core.compiler.processAnnotations=enabled
