--- android/build.properties
@@ -1,3 +1,0 @@
-titanium.platform=/home/miga/.titanium/mobilesdk/linux/6.0.0.GA/android
-android.platform=/home/miga/tools/android-sdk-linux/platforms/android-23
-google.apis=/home/miga/tools/android-sdk-linux/add-ons/addon-google_apis-google-23
