--- android/build.xml
@@ -1,56 +1,0 @@
-<project name="ti.bluetooth" default="dist">
-	<description>
-		Ant build script for Titanium Android module ti.bluetooth
-	</description>
-
-	<property name="ti.module.root" location="${basedir}"/>
-	<property file="build.properties" />
-
-	<!-- Make sure 'architectures' is in manifest file -->
-	<property file="${ti.module.root}/manifest" prefix="manifest"/>
-	<fail message="Missing required manifest key 'architectures'.">
-		<condition>
-			<not>
-				<isset property="manifest.architectures"/>
-			</not>
-		</condition>
-	</fail>
-
-	<fail message="Manifest key 'architectures' has no value.">
-		<condition>
-			<not>
-				<length string="${manifest.architectures}" trim="true" when="greater" length="0"/>
-			</not>
-		</condition>
-	</fail>
-
-	<!-- Copy documentation subdirectories -->
-	<mkdir dir="${basedir}/documentation"/>
-	<copy todir="${basedir}/documentation">
-		<fileset dir="${basedir}/../documentation"/>
-	</copy>
-
-	<!-- Copy example subdirectories -->
-	<mkdir dir="${basedir}/example"/>
-	<copy todir="${basedir}/example">
-		<fileset dir="${basedir}/../example"/>
-	</copy>
-
-	<!-- Copy assets subdirectories -->
-	<mkdir dir="${basedir}/assets"/>
-	<copy todir="${basedir}/assets">
-		<fileset dir="${basedir}/../assets"/>
-	</copy>
-
-	<!-- Copy license -->
-	<copy todir="${basedir}" file="${basedir}/../LICENSE"/>
-
-	<target name="cleancopy" description="Delete old copies">
-		<delete dir="${basedir}/documentation"/>
-		<delete dir="${basedir}/example"/>
-		<delete dir="${basedir}/assets"/>
-		<delete file="${basedir}/LICENSE"/>
-	</target>
-
-	<import file="${titanium.platform}/../module/android/build.xml"/>
-</project>
