--- android/documentation/index.md
@@ -1,39 +1,0 @@
-# ti.bluetooth Module
-
-## Description
-
-TODO: Enter your module description here
-
-## Accessing the ti.bluetooth Module
-
-To access this module from JavaScript, you would do the following:
-
-    var ti_bluetooth = require("ti.bluetooth");
-
-The ti_bluetooth variable is a reference to the Module object.
-
-## Reference
-
-TODO: If your module has an API, you should document
-the reference here.
-
-### ti_bluetooth.function
-
-TODO: This is an example of a module function.
-
-### ti_bluetooth.property
-
-TODO: This is an example of a module property.
-
-## Usage
-
-TODO: Enter your usage example here
-
-## Author
-
-TODO: Enter your author name, email and other contact
-details you want to share here.
-
-## License
-
-TODO: Enter your license/legal information here.
