--- assets/README
@@ -1,9 +1,0 @@
-Place your assets like PNG files in this directory and they will be packaged
-with your module.
-
-All JavaScript files in the assets directory are IGNORED except if you create a
-file named "ti.bluetooth.js" in this directory in which case it will be
-wrapped by native code, compiled, and used as your module. This allows you to
-run pure JavaScript modules that are pre-compiled.
-
-Note: Mobile Web does not support this assets directory.
