+++ ios/titanium.xcconfig
@@ -4,7 +4,7 @@
 // OF YOUR TITANIUM SDK YOU'RE BUILDING FOR
 //
 //
-TITANIUM_SDK_VERSION = 6.2.1.GA
+TITANIUM_SDK_VERSION = 7.5.0.GA
 
 //
 // THESE SHOULD BE OK GENERALLY AS-IS
